+++ samples/sfmtdc/test_period.cpp
@@ -18,33 +18,33 @@ int main(int argc, char * argv[])
     }
     sfmt_param params;
     char * para = argv[1];
-    params.mexp = strtoul(para, &para, 10);
+    params.mexp = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.pos1 = strtoul(para, &para, 10);
+    params.pos1 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.sl1 = strtoul(para, &para, 10);
+    params.sl1 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.sl2 = strtoul(para, &para, 10);
+    params.sl2 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.sr1 = strtoul(para, &para, 10);
+    params.sr1 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.sr2 = strtoul(para, &para, 10);
+    params.sr2 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.msk1 = strtoul(para, &para, 16);
+    params.msk1 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.msk2 = strtoul(para, &para, 16);
+    params.msk2 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.msk3 = strtoul(para, &para, 16);
+    params.msk3 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.msk4 = strtoul(para, &para, 16);
+    params.msk4 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.parity1 = strtoul(para, &para, 16);
+    params.parity1 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.parity2 = strtoul(para, &para, 16);
+    params.parity2 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.parity3 = strtoul(para, &para, 16);
+    params.parity3 = static_cast<uint32_t>(strtoul(para, &para, 16));
     para++;
-    params.parity4 = strtoul(para, &para, 16);
+    params.parity4 = static_cast<uint32_t>(strtoul(para, &para, 16));
     sfmt sf(params);
     cout << sf.getParamString() << endl;
     int se = 1;
@@ -52,7 +52,7 @@ int main(int argc, char * argv[])
     int min = INT_MAX;
     int max = -10;
     int max_weight = 0;
-    seedw.u[0] = se;
+    seedw.u[0] = static_cast<uint32_t>(se);
     sf.seed(seedw);
     min_max_poly(min, max, sf, max_weight);
     cout << "berore anni ";
@@ -86,13 +86,13 @@ void min_max_poly(int &min, int&max, sfmt& sf, int& max_weight)
     max = -10;
     for (int i = 0; i < 128; i++) {
         minpoly<w128_t>(poly, sf, i);
-        int d = deg(poly);
+        int d = static_cast<int>(deg(poly));
         if (d < min) {
             min = d;
         }
         if (d > max) {
             max = d;
-            max_weight = weight(poly);
+            max_weight = static_cast<int>(weight(poly));
         }
     }
 }
